% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-themes.R
\name{html_dependency_vmodern}
\alias{html_dependency_vmodern}
\title{vmodern dependencies}
\usage{
html_dependency_vmodern(lab_notebook = FALSE)
}
\arguments{
\item{lab_notebook}{Logical indicating whether the vmodern theme is being
used for a PCS lab notebook. If \code{TRUE}, the vmodern theme will
also load additional scripts to enable interactivity of the notebook.}
}
\description{
Dependencies on top of base material design theme.
}
\keyword{internal}
