% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scale-themes.R
\name{scale_color_vmodern}
\alias{scale_color_vmodern}
\alias{scale_colour_vmodern}
\alias{scale_fill_vmodern}
\title{vmodern color scales}
\usage{
scale_colour_vmodern(
  discrete = FALSE,
  palette = "high_contrast",
  viridis_option = "plasma",
  ...
)

scale_color_vmodern(
  discrete = FALSE,
  palette = "high_contrast",
  viridis_option = "plasma",
  ...
)

scale_fill_vmodern(
  discrete = FALSE,
  palette = "high_contrast",
  viridis_option = "plasma",
  ...
)
}
\arguments{
\item{discrete}{Generate a discrete palette? (default: \code{FALSE} -
generate continuous palette).}

\item{palette}{Name of the color palette. Must be one of "high_contrast",
"viridis", "bright_contrast", "bright_contrast2", "muted", "muted2".}

\item{viridis_option}{A character string indicating the color map option to
use for viridis. Options are "magma" (or "A"), "inferno" (or "B"),
"plasma" (or "C"), "viridis" (or "D"), "cividis" (or "E"), "rocket"
(or "F"), "mako" (or "G"), "turbo" (or "H").}

\item{...}{Additional arguments to pass to \code{ggplot2::discrete_scale()}
if \code{discrete == TRUE} and to \code{viridis::scale_colour_viridis()}
if \code{discrete == FALSE}.}
}
\description{
Modern scale functions (fill and colour/color) for \code{ggplot2}.
For \code{discrete == FALSE}, the vmodern scale uses the \code{viridis} color
palette. For \code{discrete == TRUE}, the vmodern color scale returns a
modified version of the \code{RColorBrewer} 'Dark2' theme if the number of
levels is <= 8. Otherwise, the discrete \code{viridis} color palette is used.
}
\examples{
library(ggplot2)
ggplot(iris) +
  aes(x = Sepal.Length, y = Sepal.Width, color = Petal.Length) +
  geom_point() +
  scale_color_vmodern(discrete = FALSE)
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  scale_fill_vmodern(discrete = TRUE)

}
