% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_blank}
\alias{theme_blank}
\alias{theme_blank_x}
\alias{theme_blank_y}
\alias{theme_blank_xy}
\title{Blank axis themes for ggplot.}
\usage{
theme_blank_x()

theme_blank_y()

theme_blank_xy()
}
\value{
A ggplot theme object.
}
\description{
Removes text and formatting from x and/or y-axes.
}
\examples{
require(ggplot2)

# blank x-axis theme
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  theme_blank_x()

# blank y-axis theme
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  theme_blank_y()

# blank x- and y-axes theme
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  theme_blank_xy()

}
