% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_vmodern}
\alias{theme_vmodern}
\title{A modern ggplot theme}
\usage{
theme_vmodern(
  base_family = "",
  size_preset = NULL,
  bg_color = "grey98",
  strip_bg_color = "#2c3e50",
  grid_color = "grey90",
  axis_line_width = 1,
  show_ticks = TRUE,
  x_text_angle = FALSE,
  axis_title_size = 10,
  axis_text_size = 7,
  legend_title_size = 10,
  legend_text_size = 8,
  strip_text_size = 9,
  title_size = 12,
  ...
)
}
\arguments{
\item{base_family}{Base font family.}

\item{size_preset}{One of "small", "medium", "large", "xlarge", or
"{some number of x's}large" indicating the default sizes for plot text and
titles. If \code{NULL}, defaults to values specified explicitly by
\code{axis_title_size}, \code{axis_text_size}, \code{legend_title_size},
\code{legend_text_size}, \code{strip_text_size}, \code{title_size}.}

\item{bg_color}{Color for plot background.}

\item{strip_bg_color}{Color for strip background (for
\code{ggplot2::facet_grid()} and \code{ggplot2::facet_wrap()}).}

\item{grid_color}{Color of panel grid major axes or \code{NULL} if want no
major grid lines.}

\item{axis_line_width}{Width of x and y axes lines.}

\item{show_ticks}{Logical; whether or not to show axes tick marks.}

\item{x_text_angle}{Logical; whether or not to angle x text at 45 degrees.}

\item{axis_title_size}{Font size of axis title. Ignored if \code{size_theme}
not \code{NULL}.}

\item{axis_text_size}{Font size of axis text. Ignored if \code{size_theme}
not \code{NULL}.}

\item{legend_title_size}{Font size of legend title. Ignored if
\code{size_theme} not \code{NULL}.}

\item{legend_text_size}{Font size of legend text/key. Ignored if
\code{size_theme} not \code{NULL}.}

\item{strip_text_size}{Font size of strip text. Ignored if \code{size_theme}
not \code{NULL}.}

\item{title_size}{Font size of plot title. Ignored if \code{size_theme}
not \code{NULL}.}

\item{...}{Additional arguments to pass to \code{ggplot2::theme()}}
}
\value{
A \code{ggplot2::theme()} object.
}
\description{
A modern ggplot theme
}
\examples{
require(ggplot2)
ggplot(iris) +
  aes(x = Sepal.Length, y = Sepal.Width) +
  geom_point() +
  theme_vmodern()

}
