% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scale-themes.R
\name{vmodern_pal}
\alias{vmodern_pal}
\title{vmodern color palette (discrete)}
\usage{
vmodern_pal(viridis = FALSE, option = "plasma")
}
\arguments{
\item{viridis}{Logical indicating whether or not to use the \code{viridis}
scheme.}

\item{option}{A character string indicating the color map option to use.
Eight options are available:
\itemize{
 \item "magma" (or "A")
 \item "inferno" (or "B")
 \item "plasma" (or "C")
 \item "viridis" (or "D")
 \item "cividis" (or "E")
 \item "rocket" (or "F")
 \item "mako" (or "G")
 \item "turbo" (or "H")
}}
}
\description{
The vmodern theme integrates a slight modification of the
'Dark2' color palette from \code{RColorBrewer} and the \code{viridis} palette. The
modified 'Dark2' palette is used unless the number of values exceeds 8 or
if viridis is explicitly requested.
}
\examples{
library(scales)
show_col(vmodern_pal()(8))
show_col(vmodern_pal(viridis = TRUE)(8))

}
