% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scale-themes.R
\name{vmodern_pal}
\alias{vmodern_pal}
\title{vmodern color palette (discrete)}
\usage{
vmodern_pal(palette = "high_contrast", viridis_option = "plasma")
}
\arguments{
\item{palette}{Name of the color palette. Must be one of "high_contrast",
"viridis", "bright_contrast", "bright_contrast2", "muted", "muted2".}

\item{viridis_option}{A character string indicating the color map option to
use for viridis. Options are "magma" (or "A"), "inferno" (or "B"),
"plasma" (or "C"), "viridis" (or "D"), "cividis" (or "E"), "rocket"
(or "F"), "mako" (or "G"), "turbo" (or "H").}
}
\description{
The vmodern theme integrates a slight modification of the
'Dark2' color palette from \code{RColorBrewer}, the \code{viridis} palette, and other
custom color palettes.
}
\examples{
library(scales)
show_col(vmodern_pal()(8))
show_col(vmodern_pal(palette = "viridis")(8))

}
